requirements.txt
#requirements.txt

fastapi
uvicorn
psycopg2
requests
